Refactor to move new tests into tests/functional/ to ensure they're only run for integration tests
